RUM-7082: Use dedicated semantics node mapper to resolve container frame

+++ features/dd-sdk-android-session-replay-compose/src/main/kotlin/com/datadog/android/sessionreplay/compose/internal/utils/SemanticsUtils.kt
@@ -99,8 +99,8 @@ internal class SemanticsUtils {
         }
     }
 
-    internal fun resolveBackgroundInfoId(backgroundInfo: BackgroundInfo): Long {
-        return System.identityHashCode(backgroundInfo).toLong()
+    internal fun resolveBackgroundInfoId(backgroundInfo: BackgroundInfo): Int {
+        return System.identityHashCode(backgroundInfo)
     }
 
     private fun shrinkInnerBounds(
